requirements.txt
altgraph==0.17.3
appimage-builder==1.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
contextlib2==21.6.0
decorator==5.1.1
docker==6.1.2
emrichen==0.4.0
greenlet==3.0.1
idna==3.4
jsonpath-rw==1.4.0
lief==0.13.2
numpy==1.24.3
packaging==23.1
platformdirs==3.10.0
ply==3.11
prompt-toolkit==3.0.39
psutil==5.9.6
pyaml==21.10.1
pycairo==1.24.0
PyGObject==3.44.1
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gnupg==0.5.0
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
questionary==1.10.0
requests==2.31.0
roam==0.3.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
schema==0.7.5
six==1.16.0
SQLAlchemy==2.0.23
text-unidecode==1.3
timeago==1.0.16
typing_extensions==4.8.0
Unidecode==1.3.6
urllib3==2.0.3
wcwidth==0.2.6
websocket-client==1.6.1
